[front] refactor: reuse capabilities picker list in place
